moved acknowledgement entity to acknowledgement module
