--- jslib/core/index.ts
@@ -1,2 +1,0 @@
-export * as Errors from './errors';
-export * as Http from './http';
