+++ jslib/errors/BaseError.ts
