+++ jslib/errors/HttpError.ts
