+++ jslib/errors/JsonParseError.ts
