+++ jslib/errors/NoContentError.ts
