+++ jslib/errors/ValidationError.ts
