+++ jslib/errors/index.ts
