+++ modules/acknowledgements/jsx/entities/acknowledgement/index.ts
