+++ modules/acknowledgements/jsx/entities/index.ts
